Add debug prints to xlat_tables
Defining DEBUG_XLAT_TABLE=1 will generate console output showing the
memory map and the translation table contents generated from this.

The implementation doesn't make use of printf as the stack usage of that
call path is far too high for early stages of boot where we have a tiny
coherent stack. The code also assumes that console_init() and
console_putc() are safe to use at this early stage which may not always
be the case on all platforms. It is because of this caveat, and the
amount of output generated, that we use a separate macro to enabling
this feature rather the the general DEBUG macro.

Signed-off-by: Jon Medhurst <tixy@linaro.org>