fvp: plat_io_storage: remove duplicated code
Fixes issues #41:

https://github.com/ARM-software/tf-issues/issues/41

The policy functions for each file to be loaded were implemented by
copy/pasting one method and then varying the data checked.

This patch creates a generic function to check the policy based on the
data stored in a table.

This removes the amount of duplicated code but also makes the code
simpler and more efficient.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>