TSP: Make the platform-specific makefile mandatory
The Test Secure-EL1 Payload implementation should always have a
platform-specific component.  Therefore, there should always
be a platform-specific sub-makefile for the TSP.  If there is
none then assume TSP is not supported on this specific platform
and throw an error at build time if the user tries to compile it.

Change-Id: Ibfbe6e4861cc7786a29f2fc0341035b852925193