Juno: add support for PSCI cpu_suspend api
This patch adds support for the PSCI cpu_suspend api to allow entry
into low power states until affinity level 1 i.e. cluster. It mainly
ensures that a consolidated power off command which specifies the
level to which each affinity level should be powered down is sent to
the SCP. It also ensures the MHU driver waits for the SCP to pick up a
command before doing further processing.

Change-Id: I8ad10652d388525730a228f0900a659246dd5087