Make the entry point parameter optional in load_image()
There are cases where the entry point information is useless to the
caller, e.g. when an image just needs to be loaded in memory but won't
ever be executed.

This patch allows load_image() function to take a NULL pointer as the
entry point parameter. In this case, it won't be populated.

Change-Id: Ie9394b054457706c6699926c5e0206e0c3851c56